Use obtain-auth-headers code from solid-auth-fetcher
