+++ src/index.ts
@@ -1,5 +1,4 @@
-import {getSession } from "solid-auth-fetcher";
-import {getAuthFetcher,IloginOptions} from "./obtain-auth-headers";
+import { getSession, getNodeSolidServerCookie, getAuthFetcher } from "solid-auth-fetcher";
 import fetch from "node-fetch";
 import * as SolidRest from "solid-rest";
 
@@ -10,6 +9,13 @@ let globalSession;
 
 /** PUBLIC METHODS 
  */
+export interface IloginOptions {
+  idp? : string,
+  username? : string,
+  password? : string,
+  debug? : boolean,
+  rest? : any,
+}
 export async function fetch(url:string,options:any) {
   return await globalSession.fetch(url,options);
 }
@@ -60,7 +66,8 @@ async function _getAuthSession(options:IloginOptions){
   })
 }
 async function _getAuthFetcher(options:IloginOptions,callback:Function){
-  authFetcher = await getAuthFetcher(options)
+  const cookie = await getNodeSolidServerCookie(options.idp, options.username, options.password);
+  const authFetcher = await getAuthFetcher(options.idp, cookie, "https://solid-node-client");
   let session = await getSession();
   authFetcher.onSession( async(s) => {
     let originalFetch = s.fetch;
