--- src/obtain-auth-headers.js
@@ -1,135 +1,0 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-exports.__esModule = true;
-exports.getAuthHeaders = exports.getAuthFetcher = void 0;
-var solid_auth_fetcher_1 = require("solid-auth-fetcher");
-/**
- *  this is a very slightly modified version of Michiel de Jong's
- *  obtain-auth-headers.ts from solid-crud-tests
- */
-var DEBUG = false;
-function getAuthFetcher(options) {
-    return __awaiter(this, void 0, void 0, function () {
-        var SERVER_ROOT, USERNAME, PASSWORD, authFetcher, serverLoginResult, cookie, session, redirectedTo, result;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    SERVER_ROOT = options.idp;
-                    USERNAME = options.username;
-                    PASSWORD = options.password;
-                    DEBUG = options.debug;
-                    return [4 /*yield*/, solid_auth_fetcher_1.customAuthFetcher()];
-                case 1:
-                    authFetcher = _a.sent();
-                    dlog('POSTing ' + (SERVER_ROOT + "/login/password ") + ("username=" + USERNAME + "&password=" + PASSWORD));
-                    return [4 /*yield*/, authFetcher.fetch(SERVER_ROOT + "/login/password", {
-                            headers: {
-                                "content-type": "application/x-www-form-urlencoded"
-                            },
-                            body: "username=" + USERNAME + "&password=" + PASSWORD,
-                            method: "POST",
-                            redirect: "manual"
-                        })];
-                case 2:
-                    serverLoginResult = _a.sent();
-                    cookie = serverLoginResult.headers.get('set-cookie');
-                    dlog({ cookie: cookie });
-                    return [4 /*yield*/, authFetcher.login({
-                            oidcIssuer: SERVER_ROOT,
-                            redirect: "https://solid-node-client/redirect"
-                        })];
-                case 3:
-                    session = _a.sent();
-                    dlog('got session');
-                    redirectedTo = session.neededAction.redirectUrl;
-                    _a.label = 4;
-                case 4:
-                    dlog({ redirectedTo: redirectedTo });
-                    return [4 /*yield*/, authFetcher.fetch(redirectedTo, {
-                            headers: { cookie: cookie },
-                            redirect: "manual"
-                        })];
-                case 5:
-                    result = _a.sent();
-                    redirectedTo = result.headers.get("location");
-                    if (redirectedTo === null) {
-                        throw new Error('Please add https://solid-node-client as a trusted app!');
-                    }
-                    _a.label = 6;
-                case 6:
-                    if (!(redirectedTo === null || redirectedTo === void 0 ? void 0 : redirectedTo.startsWith("https://solid-node-client"))) return [3 /*break*/, 4];
-                    _a.label = 7;
-                case 7:
-                    dlog('handling ' + redirectedTo);
-                    return [4 /*yield*/, authFetcher.handleRedirect(redirectedTo)];
-                case 8:
-                    _a.sent();
-                    return [2 /*return*/, authFetcher];
-            }
-        });
-    });
-}
-exports.getAuthFetcher = getAuthFetcher;
-// Michiel de Jong comment
-// FIXME: This is a total hack, obviously, second-guessing the
-// DI architecture of solid-auth-fetcher:
-function getAuthHeaders(urlStr, method, authFetcher) {
-    return __awaiter(this, void 0, void 0, function () {
-        var _a;
-        return __generator(this, function (_b) {
-            switch (_b.label) {
-                case 0:
-                    _a = {
-                        Authorization: JSON.parse(authFetcher.authenticatedFetcher.tokenRefresher.storageUtility.storage.map['solidAuthFetcherUser:global']).accessToken
-                    };
-                    return [4 /*yield*/, authFetcher.authenticatedFetcher.tokenRefresher.tokenRequester.dpopHeaderCreator.createHeaderToken(new URL(urlStr), method)];
-                case 1: return [2 /*return*/, (_a.DPop = _b.sent(),
-                        _a)];
-            }
-        });
-    });
-}
-exports.getAuthHeaders = getAuthHeaders;
-function dlog() {
-    var args = [];
-    for (var _i = 0; _i < arguments.length; _i++) {
-        args[_i] = arguments[_i];
-    }
-    if (DEBUG)
-        console.log(args);
-}
